lag egen eøs-mappe under behandlingsresultat for å gruppere eøs-ting
