+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/EøsPeriode/EøsPeriodeSkjema.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 import { Fieldset } from '@navikt/ds-react';
 import type { Felt } from '@navikt/familie-skjema';
 
-import MånedÅrVelger from '../../../../../../komponenter/MånedÅrInput/MånedÅrVelger';
-import type { IIsoMånedPeriode } from '../../../../../../utils/dato';
+import MånedÅrVelger from '../../../../../../../komponenter/MånedÅrInput/MånedÅrVelger';
+import type { IIsoMånedPeriode } from '../../../../../../../utils/dato';
 
 const FlexDiv = styled.div`
     width: 32rem;
