+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/Kompetanse/KompetanseSkjema.tsx
@@ -5,9 +5,8 @@ import styled from 'styled-components';
 import { Alert, Heading, Table } from '@navikt/ds-react';
 
 import KompetanseTabellRad from './KompetanseTabellRad';
-import type { IBehandling } from '../../../../../../typer/behandling';
-import type { IRestKompetanse } from '../../../../../../typer/eøsPerioder';
-import { EøsPeriodeStatus } from '../../../../../../typer/eøsPerioder';
+import type { IBehandling } from '../../../../../../../typer/behandling';
+import { EøsPeriodeStatus, type IRestKompetanse } from '../../../../../../../typer/eøsPerioder';
 
 const KompetanseContainer = styled.div`
     margin-top: 5rem;
